+++ adivinaLaPalabra/src/test/java/com/soprasteria/adivinaLaPalabra/controller/RoundControllerTest.java
@@ -1,8 +1,10 @@
 package com.soprasteria.adivinaLaPalabra.controller;
 
 import com.soprasteria.adivinaLaPalabra.dto.RoundResponse;
+import com.soprasteria.adivinaLaPalabra.dto.WordResponse;
 import com.soprasteria.adivinaLaPalabra.service.RoundService;
 import com.soprasteria.adivinaLaPalabra.service.RoundServiceImpl;
+import com.soprasteria.adivinaLaPalabra.service.WordExistServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -12,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 
@@ -20,6 +23,9 @@ class RoundControllerTest {
 
     @Mock
     private RoundServiceImpl roundService;
+
+    @Mock
+    private WordExistServiceImpl wordExistService;
 
     @InjectMocks
     private RoundController roundController;
@@ -44,6 +50,28 @@ class RoundControllerTest {
                 new ResponseEntity<>(null, HttpStatus.UNPROCESSABLE_ENTITY);
 
         assertEquals(responseEntityExpected, roundController.newRound());
+    }
+
+    @Test
+    void returnFalseIsNotWord() {
+        final String word = "word";
+        WordResponse wordResponse = new WordResponse(false);
+        when(wordExistService.checkWord(any())).thenReturn(wordResponse);
+
+        ResponseEntity<WordResponse> responseEntityExpected = new ResponseEntity<>(wordResponse, HttpStatus.NOT_FOUND);
+
+        assertEquals(responseEntityExpected, roundController.wordExist(word, any()));
+    }
+
+    @Test
+    void returnTrueIsWord() {
+        final String word = "word";
+        WordResponse wordResponse = new WordResponse(true);
+        when(wordExistService.checkWord(any())).thenReturn(wordResponse);
+
+        ResponseEntity<WordResponse> responseEntityExpected = ResponseEntity.ok(wordResponse);
+
+        assertEquals(responseEntityExpected, roundController.wordExist(word, any()));
     }
 
 }
